# specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users": {
      "get": {
        "operationId": "AdminUsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "jane",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Role"
            }
          },
          {
            "name": "isDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft-deleted users",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "createdAfter",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2024-01-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "createdBefore",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "example": "2024-12-31T23:59:59Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAdminUserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List all users with filters",
        "tags": [
          "Admin – Users"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get a user by ID (admin view)",
        "tags": [
          "Admin – Users"
        ]
      },
      "patch": {
        "operationId": "AdminUsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update a user (role, isActive, token invalidation)",
        "tags": [
          "Admin – Users"
        ]
      },
      "delete": {
        "operationId": "AdminUsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Soft-delete a user",
        "tags": [
          "Admin – Users"
        ]
      }
    },
...
    "/products/{productId}/reviews": {
      "get": {
        "operationId": "ReviewsController_findByProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReviewResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "List approved reviews for a product",
        "tags": [
          "Reviews"
        ]
      },
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Submit a review for a product",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/{id}": {
      "delete": {
        "operationId": "ReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Review UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Not the review owner"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete your own review",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2Shop API",
    "description": "REST API for the O2Shop e-commerce platform. Public endpoints need no authentication. Customer endpoints require a JWT bearer token (from POST /auth/login). Admin endpoints additionally require the `admin` role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access_token value from POST /auth/login."
      }
    },
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "regular",
          "admin"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "example": [
              "email must be an email"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "AdminUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "displayName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "googleLinked": {
            "type": "boolean",
            "example": false
          },
          "role": {
            "example": "regular",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "tokenVersion": {
            "type": "number",
            "example": 1
          },
          "deletedAt": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-06-01T00:00:00.000Z"
          },
          "version": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "id",
          "email",
          "googleLinked",
          "role",
          "isActive",
          "createdAt",
          "tokenVersion",
          "updatedAt",
          "version"
        ]
      },
      "PaginatedAdminUserResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserResponseDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ]
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "example": "regular",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "resetTokenVersion": {
            "type": "boolean",
            "description": "Set to true to invalidate all existing tokens for this user",
            "example": false
          }
        }
      },
...
      "ReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Jane D."
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "example": 8
          },
          "content": {
            "type": "string",
            "example": "Great quality, fast shipping!"
          },
          "photoUrls": {
            "example": [
              "https://example.com/photo.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "displayName",
          "rating",
          "content",
          "createdAt"
        ]
      },
      "PaginatedReviewResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponseDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "displayName": {
            "type": "string",
            "example": "Jane D."
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "example": 8
          },
          "content": {
            "type": "string",
            "example": "Great quality, fast shipping!"
          },
          "photoUrls": {
            "example": [
              "https://example.com/photo.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "displayName",
          "rating",
          "content"
        ]
      },
...
    }
  }
}